+++ l2/agentPlannerL2/agentPlannerL2.ts
@@ -24,6 +24,7 @@ import '/_102020_/l2/agentPlannerL2/steps/entry10/agentP2Entry.js';
 import '/_102020_/l2/agentPlannerL2/steps/workspaces20/agentP2Workspaces.js';
 import '/_102020_/l2/agentPlannerL2/steps/contracts30/agentP2Contracts.js';
 import '/_102020_/l2/agentPlannerL2/steps/shared40/agentP2Shared.js';
+import '/_102020_/l2/agentPlannerL2/steps/requests50/agentP2Requests.js';
 
 export function createAgent(): IAgentAsync {
   return {